Reproduced the circuit diagram
This is based on pictures from Nik's original diagram, and the
breadboard prototype at
https://nikharris.com/usbremote-an-open-source-infrared-remote-reciever/

Note that there was at least one mistake in Nik's original diagram that
was not present in his breadboard prototype photo. In the diagram the
pulldown resistor was between the reciever's data pin and ATMega328's
pin 14. This might've still worked, but I believe it was intended to be
a pulldown as it is in the breadboard